requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.34.0
numpy>=1.24.0
soundfile>=0.12.1
scipy>=1.11.0
